requirements.txt
comtypes==1.4.12
et_xmlfile==2.0.0
ImageHash==4.3.2
imageio==2.37.0
lazy_loader==0.4
lxml==5.4.0
MouseInfo==0.1.3
networkx==3.2.1
numpy==2.0.2
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pillow==11.2.1
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==2.0.1
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytweening==1.2.0
pytz==2025.2
PyWavelets==1.6.0
pywin32==311
pywinauto==0.6.9
scikit-image==0.24.0
scipy==1.13.1
six==1.17.0
tifffile==2024.8.30
typing_extensions==4.14.0
tzdata==2025.2
xlsxwriter==3.2.9
